Day5.part1(): Eliminate unnecessary ArrayList structures.
The previous version of part1() scans the ArrayList that contains the full list of coordinate records, builds up two new ArrayLists (one each for horizontal line records and vertical line records), and after the initial scan of the full ArrayList it reads through these new ArrayLists and draws the lines onto the grid.

This new version moves the line-drawing code into the initial scan of the full list of coordinate records, drawing each line onto the grid as we encounter it in the master list, instead of building up extra ArrayLists and saving them for later. The improvement here is to just do less work; the end result is identical and the new version uses a bit less time and space to do it.